requirements.txt
typing-extensions
langchain-experimental
pandas
yfinance
praw
feedparser
stockstats
eodhd
IPython
langgraph
chromadb
setuptools
backtrader
akshare
tushare
finnhub-python
parsel
requests
tqdm
pytz
redis
chainlit
rich
questionary
langchain-google-genai
alpaca-trade-api
